Leaderboard
Here is an unordered list of players with information about the number of points scored in a certain computer game. I build a leaderboard implementing the sorting algorithm using a function that is called from main(). The sorting function take the input array and the maximum size of the leaderboard as an argument and return a new array with the leaderboard.