Cheapest Way To Travel
You are given a N×M
map(grid). An agent wants to travel from the start (top left) of the map to the finish line(bottom right).
Each cell on the map, contains a non-negative number, denoting the cost of passing that cell.
Your agent can only move one step at each time only to one of these direction — Right or down. For moving to a direction, a cell on that direction must be available on the map.
Return a path for the agent to travel from start to end which cost the agent the least cost.