Anonymous Alphabet
An unknown civilization uses a unique alphabet. You will be given a dictionary of words from their language. The words are sorted lexicographically according to the rules of their alphabet. Your task is to deduce the correct order of characters in their alphabet.

If multiple solutions exist, return any of them. If it is impossible to determine a valid alphabet order, output "Doh".